ai_scientist/requirements.txt
groq
python-dotenv
streamlit
optuna
scikit-learn
xgboost
pandas
numpy
fpdf2
